move entities to designated folder
