+++ internal/cafeteria/entities/chatgpt-message.go
@@ -1,4 +1,4 @@
-package cafeteria
+package entities
 
 type Message struct {
 	Role    string `json:"role"`
