+++ internal/cafeteria/entities/chatgpt-response.go
@@ -1,4 +1,4 @@
-package cafeteria
+package entities
 
 type GPTResponse struct {
 	Message *Message `json:"message"`
